refactor: rename class to more meaningful name
