+++ app/src/main/java/com/example/playlistmaker/domain/api/LocalDataSource.kt
@@ -1,6 +1,6 @@
 package com.example.playlistmaker.domain.api
 
-interface LocalRepository {
+interface LocalDataSource {
     fun getBoolean(key: String, defValue: Boolean): Boolean
     fun putBoolean(key: String, value: Boolean)
     fun getString(key: String, defValue: String?): String?
